Added the ability to generate outlines for PDFs
--outline-tags allows to specify the HTML tags which should be
considered for the outline. The tags are expected to be given in
order of hierachy, for example, 'h1,h2' will trigger a generation
with h1 elements as top level outline entries and h2 as their
childs.

Ideally this would not be required if Chromium would add
this directly. So if these bugs are closed this can probably be
removed again:
- https://bugs.chromium.org/p/chromium/issues/detail?id=840455
- https://github.com/GoogleChrome/puppeteer/issues/1778

This code is heavily based on @Hopding's comment at:
https://github.com/Hopding/pdf-lib/issues/127#issuecomment-502450179